test/common: check if pid is alive before kill
Need to check if process still exists before killing it.
Otherwise if process crashed and no longer exists kill will
return with "no such process" error and prevent other steps
in trap from executing.

Reviewed-on: https://review.gerrithub.io/425353
Chandler-Test-Pool: SPDK Automated Test System <sys_sgsw@intel.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>

+++ test/common/autotest_common.sh
@@ -338,9 +338,11 @@ function killprocess() {
 		exit 1
 	fi
 
-	echo "killing process with pid $1"
-	kill $1
-	wait $1
+	if kill -0 $1; then
+		echo "killing process with pid $1"
+		kill $1
+		wait $1
+	fi
 }
 
 function iscsicleanup() {
